refactor: replace Plausible initialization with startPlausible function
